requirements.txt
fastapi
uvicorn
openai
fpdf
python-dotenv
